icmpv6 client ndp/echo split done
